Project/Data-Engineering/Polars/Project2/requirements.txt
polars
pydantic
sqlalchemy
psycopg2-binary
elasticsearch==7.13.0
apache-airflow
python-dotenv